0.7.12.pcl-build1.1: 	So, I hope this goes onto the right branch.
	Begin (or continue, depending on how you look at it) motion of
		PCL build into the main build:
	... hack the package setup logic to include shadows;
	... export all the shadowed symbols from SB!PCL, otherwise
		genesis drops the information on the floor;
	... migrate compiler-support, macros and early-low into the main
		build;

	Also a fix for a stupid bug, that might want to be merged before
		the main thrust of this branch.
	... (documentation nil t) no longer talks about
		SB-SLOT-ACCESSOR-NAME :-)

The thrust of this branch is to rationalize the build process of PCL.
This means that in scope is the removal of the bootstrap process (NIL
and :BRAID *BOOT-STATE*s) on the basis that instead we can arrange that
our first-stage (cross-) compiler knows enough to be able to compile the
target PCL, without having to execute any target PCL code.  Out of
scope, unless it turns out to be really easy and obvious, is fixing the
CL:CLASS/SB-PCL:CLASS distinction, though it is to be hoped that this
effort into rationalizing the build will make fixing the breakage more
obvious, either by virtue of code simplification or by virtue of having
read all the code at least once.

While it will be attempted to pass all tests at each CVS checkin, it's
entirely possible that manipulating the build will get hairy, and a
checkin will want to be made at an intermediate stage; therefore, this
work is done on a branch, with tag pcl_build_1_branch (no, I hope there
won't be a pcl_build_2_branch, but one never knows...)