0.7.12.pcl-build1.5         Move defclass and defs into cold build         ... define COLD-DEFCLASS, and use it to make early classes         ... simplify DEFCLASS and s/REAL-LOAD-DEFCLASS/LOAD-DEFCLASS         ... only do FBOUNDP checks when we're in charge of what a                 function name is
This is perhaps indicative of the way to go.  We'll use COLD-DEFFOO to
build up early classes/methods/generic functions, which will then be
turned into late ones in early COLD-INIT, so that by the time
ordinary top-level forms run we're all set up and ready to cope with
LOAD-DEFCLASS requests.

Hrm.  I see in our cold-init sequence !CLASS-FINALIZE, which would seem
to break this plan.  Maybe !CLASS-FINALIZE will need to run earlier --
pre-genesis -- so that Lisp classes are actually finalized before they
currently are?  Or maybe it's a no-op anyway?  Currently I believe that
there are no classes in *FORWARD-REFERENCED-LAYOUTS* that don't have a
layout.