0.7.12.pcl-build1.6: 	And so falls src/pcl/cache.lisp 	... nothing really difficult in there 	... and bye bye placeholder SB!PCL::CHECK-WRAPPER-VALIDITY
OK, but now it gets difficult.  The next files, dlisp and dlisp2,
involve references to #'(sb-kernel:instance-lambda ...), which isn't a
very happy concept on the cross-compilation host.  Sadly, just putting
them into :not-host mode doesn't work either, as there are uses of
these instance-lambda-making macros in dlisp2, so they need to be
defined in some sense.  Ugh.  One to sleep on, I suspect.