0.8.0.78.vector-nil-string.6: 	Cosmetic cleanups 	... invert the test in %SXHASH-SIMPLE-SUBSTRING, since that 		way it stands a chance of working even when putative 		(VECTOR CHARACTER) != (VECTOR BASE-CHAR) 	... some reasonable slight changes to RUN-PROGRAM / Unix 		environment stuff.
The ALIEN/C-STRING issue looks like it's not a problem; all of the work
there is explicitly done in terms of BASE-STRINGs.  I don't think anyone
is likely to complain at a type error should they attempt to pass an
object of (VECTOR NIL 0) to foreign lands as the null string, and even
if they do I feel unmotivated to care without sufficient inducement. :-)