1.0.27.35: minimize the potential for presenting hosts with choices
By having minimal debug names for toplevel forms and component names, we
avoid having arbitrary gensyms or, horror of horror, QUOTE: which is
printed differently in different implementations...

2 commit messages follow:

minimal debug names for cross-compiled top-level forms

Otherwise we run the risk of getting arbitrary gensyms dumped as part of
the debug name.

bandage for ' vs QUOTE in two files

Make FIND-COMPONENT-NAME in the XC (which names components, whose names
are dumped in xc fasls) use only the first symbol in the context.  That
will be generally lame but avoids any current instances of QUOTE, which
prints differently in different implementations when pretty-printing is
off.