1.0.28.31: remove :PREFIX/:SUFFIX from DEFENUM for MORE GREPPABILITY
Searching for useful things like even-fixnum-lowtag in the source won't
find the actual definition due to cleverness with DEFENUM.  This cleverness
is the source of several FIXMEs throughout the code.  I can personally
remember being baffled at the magic of even-fixnum-lowtag and friends
magically springing into existence when I could not easily grep for their
definitions.  In the interest of clarity, remove the cleverness and update
all relevent DEFENUMs.

In the interest of OAOO, move TRACE-TABLE-* constants to generic code while
we're at it, too.  Doing the same thing for the *-TRAP constants is left
as an exercise for another day.