1.0.32.5: defend against full MAKE-ARRAY before the type system is ready
This is basically defence against bugs of the form lp #316323, where the
wrong answer comes out of a make-array before SUBTYPEP is working
properly.  We should be able to arrange the build so that we never have
to do a full MAKE-ARRAY before we're ready for it.

Slight modifications of a couple of macros: VOP-related and multibyte
encoding-related.  There should be no user-visible change.