1.0.32.12: Fix slot-value on specialized parameters in SVUC methods
The slot accessors' functions were being computed too early, requiring a
wrapper (for the PV optimization) when none was available.  The fix
delays the computation by the usual trick of installing a closure that
will perform the slot accessor computation on demand.

Include a correct MOP test for this (slightly hard because of
constraints about order of instantiation: see lp #473699, reported by
Lars Rune Nodstal).