1.0.33.5: improve DESCRIBE (#lp488976)
Instead of printing the object type and address in an around method on
DESCRIBE-OBJECT, make it a normal method that all other methods are
supposed to call via CALL-NEXT-METHOD if they wish. This makes it
possible to completely change the output of DESCRIBE for one class
while leaving other classes alone.

Also, kill the unused *IN-DESCRIBE* var in the process.