1.0.37.13: Make SIGNAL-SEMAPHORE interruptable.
Although a SEMAPHORE's mutex is private to that SEMAPHORE, and the
mutex is released during the wait in WAIT-ON-SEMAPHORE, it's still
possible to starve in SIGNAL-SEMAPHORE while waiting to acquire the
mutex because a deadline handler around WAIT-ON-SEMAPHORE could be
running in the implicitly called CONDITION-WAIT which reacquires the
mutex.

Hence make sure that the call to GET-MUTEX in SIGNAL-SEMAPHORE is
interruptable.