1.0.43.42: Constant fold IFs before performing IF/IF conversion
IF/IF-conversion is somewhat similar to "the trick": when a conditional
branches on an lvar that may be written to by multiple nodes, duplicate
the conditional after each of the writers.

We used to perform that before constant-folding IFs.  We now do the
reverse: it's always better to completely flush the branch away than to
duplicate it.