Ignore a known failure in script.test.sh on Darwin
script.test.sh includes a test case that fails on Darwin because of
a bug in SBCL.  However, this is the first release to run this test at
all.  Lacking a proper mechanism to mark the test case as a known
failure on Darwin, skip it entirely on that platform for now.

Thanks to Fare for testing.