arch-assem.S: Update fun-end breakpoint magic for pickier make-lisp-obj.
  * We're about to cause SB-KERNEL:MAKE-LISP-OBJ to use the same
validation on CHENEYGC as it always has on GENCGC.

  * In order to not break function-end breakpoints, update them
all now to have the same header structure as on PPC (which had
to be made to work on GENCGC some time back).

  * While we're here, add a comment to each copy pointing to the
PPC version as having an actual explanation for the magic.

  * And update the PPC version commentary to the soon-to-be
reality of MAKE-LISP-OBJ always verifying pointers, instead of
only verifying pointers on GENCGC.

  * Finally, remove the no-longer-necessary code in debug-int
that set the header value, leaving a historic-information
comment behind.